Fix the printf: not enough arguments bug
By sourcing the segments and the events in the function, we lose the
variables defined in them after the function ends because of fish’s
scope rules.
This fix disables the scope shadowing of the function so the variables
defined in the segments and in the events can be accessed outside of
the function.